Implemented RoleService and PersonController, implemented convertPeriod, convertDate in util

+++ src/main/java/com/project/shareholder/service/StageServiceImpl.java
@@ -10,6 +10,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.time.YearMonth;
 import java.util.List;
 import java.util.UUID;
 
@@ -22,7 +23,6 @@ public class StageServiceImpl implements StageService {
     @Override
     public Stage create(StageRequest stageRequest) throws DatabaseException {
         Stage stage = new Stage();
-        stage.setId(UUID.fromString(stageRequest.getId()));
         stage.setName(stageRequest.getName());
         stage.setQuantity(stageRequest.getQuantity());
         stage.setDateStartedAt(stageRequest.getDateStartedAt());
@@ -56,10 +56,14 @@ public class StageServiceImpl implements StageService {
     }
 
     @Override
-    public Stage deactivate(StageRequest stageRequest) throws DatabaseException {
-        Stage stage = new Stage();
-        stage.setId(UUID.fromString(stageRequest.getId()));
+    public Stage deactivate(String id) throws DatabaseException {
+        Stage stage;
         try {
+            stage = stageDao.retrieveById(UUID.fromString(id));
+            if (null == stage) {
+                throw new NotFoundException(Constants.NOT_FOUND_MESSAGE);
+            }
+
             stageDao.deactivateObj(stage);
         } catch (Exception exception) {
             throw new DatabaseException(Constants.DATABASE_MESSAGE);
@@ -69,21 +73,30 @@ public class StageServiceImpl implements StageService {
     }
 
     @Override
-    public Stage delete(StageRequest stageRequest) throws DatabaseException {
-        Stage stage = new Stage();
-        stage.setId(UUID.fromString(stageRequest.getId()));
+    public String delete(String id) throws DatabaseException {
+        Stage stage;
         try {
+            stage = stageDao.retrieveById(UUID.fromString(id));
+            if (null == stage) {
+                throw new NotFoundException(Constants.NOT_FOUND_MESSAGE);
+            }
+
             stageDao.deleteObj(stage);
         } catch (Exception exception) {
             throw new DatabaseException(Constants.DATABASE_MESSAGE);
         }
 
-        return stage;
+        return id;
     }
 
     @Override
-    public Stage retrieveById(StageRequest stageRequest) throws NotFoundException {
-        return stageDao.retrieveById(UUID.fromString(stageRequest.getId()));
+    public Stage retrieveById(String id) throws NotFoundException {
+        return stageDao.retrieveById(UUID.fromString(id));
+    }
+
+    @Override
+    public Stage retrieveByPeriod(YearMonth period) throws NotFoundException {
+        return stageDao.retrieveByPeriod(period);
     }
 
     @Override
